+++ packages/vendure-plugin-metrics/src/metrics.plugin.ts
@@ -11,7 +11,15 @@ import { AverageOrderValueMetric } from './api/metrics/average-order-value';
 import { UnitsSoldMetric } from './api/metrics/units-sold-metric';
 
 export interface MetricsPluginOptions {
+  /**
+   * The enabled metrics shown in the widget.
+   */
   metrics: MetricStrategy<any>[];
+  /**
+   * The number of past months to display in the metrics widget.
+   * If your shop has a lot of orders, consider using only the last 3 months for example.
+   */
+  displayPastMonths: number;
 }
 
 @VendurePlugin({
@@ -33,10 +41,14 @@ export class MetricsPlugin {
       new AverageOrderValueMetric(),
       new UnitsSoldMetric(),
     ],
+    displayPastMonths: 14,
   };
 
-  static init(options: MetricsPluginOptions): typeof MetricsPlugin {
-    this.options = options;
+  static init(options: Partial<MetricsPluginOptions>): typeof MetricsPlugin {
+    this.options = {
+      ...this.options,
+      ...options,
+    };
     return MetricsPlugin;
   }