+++ packages/vendure-plugin-metrics/CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.6.0 (2024-11-19)
 
+- Allow showing metrics of past X months instead of always past 12 months.
 - Removed conversion metric, as it is can not be accurately calculated based on created orders alone
 - Small improvement in query performance